Fix type symbol acquiring
When the node that the schema is being generated from is a type, `type.symbol` is `__type`, and `type.aliasSymbol` cointains the actual type name.

This commit changes the code so we consider `types.aliasSymbol` first and, if that does not exist (which is the case for interfaces), then we use `type.symbol`.